pyproject.toml
[tool.poetry]
name = "blueprint_api"
version = "0.1.0"
description = ""
authors = [
    "Tom Camp <tom.camp@civicactions.com>"
]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.12.0"
whitenoise = "^6.1.0"
django-guardian = "^2.4.0"
pathlib = "^1.0.1"
jsonschema = "^4.14.0"
django-filter = "^22.1"
pydantic = "^1.9.2"
requests = "^2.28.1"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.3"

[tool.poetry.dev-dependencies]
pylint = "^2.14.5"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Pylint config
[tool.pylint.main]
recursive = true
load-plugins = ["pylint_django"]
ignore = ["migrations/*", "manage.py"]

[tool.pylint.format]
max-line-length = 120


[tool.pylint."messages control"]
disable = ["missing-module-docstring", "missing-class-docstring", "missing-function-docstring", "django-not-configured", "unspecified-encoding", "inconsistent-return-statements", "duplicate-code", "import-outside-toplevel"]